pkg/load/configAgent: make some functions more generic
